renamse firebase server file
