+++ src/lib/server/Feature/Auth/idManager.ts
@@ -1,5 +1,5 @@
 import { UserId } from '$lib/client/Domain/ValueObjects/BookInfo/UserId';
-import { firebaseAdminAuth } from '$lib/server/Feature/Auth/firebase.server';
+import { firebaseAdminAuth } from '$lib/server/Feature/Auth/firebase';
 import { redirect } from '@sveltejs/kit';
 
 /**
